Update imports for gym_unity
